+++ ebs-20210730/ChangeLog.md
@@ -1,3 +1,11 @@
+2022-07-15 Version: 1.3.5
+- Add describe dbsc for ecs api.
+- Add disk replica pair api.
+- Add disk replica group api.
+- Bug fix.
+- SourceRegionId change to RegionId.
+- Add Tag api.
+
 2022-05-27 Version: 1.2.1
 - Add describe dbsc for ecs api.
 - Add disk replica pair api.